Create Helm chart repository in the project
Users will be able to pull the chart from the repository directly
instead of being required to download the complete repository just to
install the chart.

The GitHub action introduced here requires "gh-pages" branch to exist in
the repository. If the branch does not exist, the action will fail.
Further, this branch must be designated in the repository's settings to
provide the source for GitHub Pages.

This new branch will contain only the chart and possibly a README.md
file but not any code. The optional README.md file is not the same as
the one from the master branch. While there is nothing to prevent these
two files to be the same, that is not practical since they would have to
be kept in sync manually. It is a better idea to create a Helm related
README.md that may point to the README.md from the master branch if
necessary. Because of this, it makes sense to create gh-pages as an
orphaned branch.

For this action to work correctly, one must update the version number in
deploy/dynu-webhook/Chart.yaml, otherwise it will fail to create a new
release because of the name collision.

For more information, refer to

https://github.com/helm/chart-releaser-action