Get domain from cert-manager
The webhook needs the Dynu's domain id to make the API calls that manage
the TXT records. It determines the id by matching the zone coming from
the static configuration and and an item from the list returned from an
API. It then combines the zone and FQDN received from cert-manager
through a regular expression to create a value for the TXT record.
However, it is simpler to call a different Dynu API that can do both of
these tasks at once.

This change makes passing the zone from a separate configuration
redundant.